Removed: duplicate method declaration for getting time in millis
