+++ src/store/mdmConfig.ts
@@ -1,6 +1,6 @@
 import logger from "@/logger";
+import { getTimeInMillis } from "@/utils";
 import { api } from "@common";
-import { DateTime } from "luxon";
 import { defineStore } from "pinia";
 
 export const useMdmConfigStore = defineStore("mdmConfig", {
@@ -178,14 +178,6 @@ export const useMdmConfigStore = defineStore("mdmConfig", {
     },
     async fetchGlobalStats() {
       const moquiStatuses = "DmlsCancelled,DmlsCrashed,DmlsFailed,DmlsFinished,DmlsPending,DmlsQueued,DmlsRunning"
-      const getTimeInMillis = (value: any) => {
-        if (!value) return 0;
-        if (typeof value === "number") return value;
-        const isoDateTime = DateTime.fromISO(value);
-        if (isoDateTime.isValid) return isoDateTime.toMillis();
-        const sqlDateTime = DateTime.fromSQL(value);
-        return sqlDateTime.isValid ? sqlDateTime.toMillis() : 0;
-      };
 
       const getAverageProcessingTime = (logs: Array<any>) => {
         const finishedLogs = logs.filter((log: any) => log.createdDate && (log.finishDateTime || log.lastUpdatedTxStamp));
